+++ src/components/NodeModal/index.tsx
@@ -1,0 +1,71 @@
+import { MantineProvider, Modal, Stack, Tabs } from '@mantine/core'
+
+import { HTTPForm } from './HTTPForm'
+import { SSForm } from './SSForm'
+import { SSRForm } from './SSRForm'
+import { Socks5Form } from './Socks5Form'
+import { TrojanForm } from './TrojanForm'
+import { V2rayForm } from './V2rayForm'
+
+export const NodeModal = ({ opened, onClose }: { opened: boolean; onClose: () => void }) => {
+  return (
+    <Modal opened={opened} onClose={onClose} title="Configure Node" size="md">
+      <MantineProvider
+        theme={{
+          components: {
+            TabsPanel: { defaultProps: { pt: 'md' } },
+            Stack: { defaultProps: { spacing: 'sm' } },
+          },
+        }}
+        inherit
+      >
+        <Tabs defaultValue="v2ray" pt="xs">
+          <Tabs.List position="center">
+            <Tabs.Tab value="v2ray">V2RAY</Tabs.Tab>
+            <Tabs.Tab value="ss">SS</Tabs.Tab>
+            <Tabs.Tab value="ssr">SSR</Tabs.Tab>
+            <Tabs.Tab value="trojan">Trojan</Tabs.Tab>
+            <Tabs.Tab value="http">HTTP</Tabs.Tab>
+            <Tabs.Tab value="socks5">SOCKS5</Tabs.Tab>
+          </Tabs.List>
+
+          <Tabs.Panel value="v2ray">
+            <Stack>
+              <V2rayForm />
+            </Stack>
+          </Tabs.Panel>
+
+          <Tabs.Panel value="ss">
+            <Stack>
+              <SSForm />
+            </Stack>
+          </Tabs.Panel>
+
+          <Tabs.Panel value="ssr">
+            <Stack>
+              <SSRForm />
+            </Stack>
+          </Tabs.Panel>
+
+          <Tabs.Panel value="trojan">
+            <Stack>
+              <TrojanForm />
+            </Stack>
+          </Tabs.Panel>
+
+          <Tabs.Panel value="http">
+            <Stack>
+              <HTTPForm />
+            </Stack>
+          </Tabs.Panel>
+
+          <Tabs.Panel value="socks5">
+            <Stack>
+              <Socks5Form />
+            </Stack>
+          </Tabs.Panel>
+        </Tabs>
+      </MantineProvider>
+    </Modal>
+  )
+}
